apps/backend/requirements.txt
fastapi
uvicorn[standard]
google-adk
python-dotenv
sqlalchemy[asyncio]
asyncpg
pydantic-settings